Centralize core settings
* Move all SettingKey to settings.py
* Alphabetize SettingKey definitions
* Move all SettingDefault to settings.py
* Alphabetize SettingDefault definitions
* Move setting_utilities.default functions under SettingDefault as private
  staticmethods
* Move setting_utilities.validator functions under a SettingValidator class
  as private staticmethods
* Ensure ValidationExcepton is called with a second argument in validators
* Document setting constant changes
