+++ girder/constants.py
@@ -123,82 +123,6 @@ class AccessType(object):
             raise ValueError('Invalid AccessType: %d.' % level)
 
 
-class SettingKey(object):
-    """
-    Core settings should be enumerated here by a set of constants corresponding
-    to sensible strings.
-    """
-    ADD_TO_GROUP_POLICY = 'core.add_to_group_policy'
-    API_KEYS = 'core.api_keys'
-    BANNER_COLOR = 'core.banner_color'
-    BRAND_NAME = 'core.brand_name'
-    COLLECTION_CREATE_POLICY = 'core.collection_create_policy'
-    PRIVACY_NOTICE = 'core.privacy_notice'
-    CONTACT_EMAIL_ADDRESS = 'core.contact_email_address'
-    COOKIE_LIFETIME = 'core.cookie_lifetime'
-    CORS_ALLOW_HEADERS = 'core.cors.allow_headers'
-    CORS_ALLOW_METHODS = 'core.cors.allow_methods'
-    CORS_ALLOW_ORIGIN = 'core.cors.allow_origin'
-    CORS_EXPOSE_HEADERS = 'core.cors.expose_headers'
-    EMAIL_FROM_ADDRESS = 'core.email_from_address'
-    EMAIL_HOST = 'core.email_host'
-    EMAIL_VERIFICATION = 'core.email_verification'
-    ENABLE_PASSWORD_LOGIN = 'core.enable_password_login'
-    GIRDER_MOUNT_INFORMATION = 'core.girder_mount_information'
-    ENABLE_NOTIFICATION_STREAM = 'core.enable_notification_stream'
-    REGISTRATION_POLICY = 'core.registration_policy'
-    ROUTE_TABLE = 'core.route_table'
-    SECURE_COOKIE = 'core.secure_cookie'
-    SERVER_ROOT = 'core.server_root'
-    SMTP_ENCRYPTION = 'core.smtp.encryption'
-    SMTP_HOST = 'core.smtp_host'
-    SMTP_PASSWORD = 'core.smtp.password'
-    SMTP_PORT = 'core.smtp.port'
-    SMTP_USERNAME = 'core.smtp.username'
-    UPLOAD_MINIMUM_CHUNK_SIZE = 'core.upload_minimum_chunk_size'
-    USER_DEFAULT_FOLDERS = 'core.user_default_folders'
-
-
-class SettingDefault(object):
-    """
-    Core settings that have a default should be enumerated here with the
-    SettingKey.
-    """
-    defaults = {
-        SettingKey.ADD_TO_GROUP_POLICY: 'never',
-        SettingKey.API_KEYS: True,
-        SettingKey.BANNER_COLOR: '#3F3B3B',
-        SettingKey.BRAND_NAME: 'Girder',
-        SettingKey.COLLECTION_CREATE_POLICY: {
-            'open': False,
-            'groups': [],
-            'users': []
-        },
-        SettingKey.CONTACT_EMAIL_ADDRESS: 'kitware@kitware.com',
-        SettingKey.PRIVACY_NOTICE: 'https://www.kitware.com/privacy',
-        SettingKey.COOKIE_LIFETIME: 180,
-        # These headers are necessary to allow the web server to work with just
-        # changes to the CORS origin
-        SettingKey.CORS_ALLOW_HEADERS:
-            'Accept-Encoding, Authorization, Content-Disposition, '
-            'Content-Type, Cookie, Girder-Authorization, Girder-OTP, Girder-Token',
-        SettingKey.CORS_EXPOSE_HEADERS: 'Girder-Total-Count',
-        # An apache server using reverse proxy would also need
-        #  X-Requested-With, X-Forwarded-Server, X-Forwarded-For,
-        #  X-Forwarded-Host, Remote-Addr
-        SettingKey.EMAIL_VERIFICATION: 'disabled',
-        SettingKey.EMAIL_FROM_ADDRESS: 'Girder <no-reply@girder.org>',
-        SettingKey.ENABLE_PASSWORD_LOGIN: True,
-        SettingKey.ENABLE_NOTIFICATION_STREAM: True,
-        SettingKey.REGISTRATION_POLICY: 'open',
-        SettingKey.SMTP_HOST: 'localhost',
-        SettingKey.SMTP_PORT: 25,
-        SettingKey.SMTP_ENCRYPTION: 'none',
-        SettingKey.UPLOAD_MINIMUM_CHUNK_SIZE: 1024 * 1024 * 5,
-        SettingKey.USER_DEFAULT_FOLDERS: 'public_private'
-    }
-
-
 class SortDir(object):
     ASCENDING = 1
     DESCENDING = -1
