+++ girder/models/setting.py
@@ -1,17 +1,12 @@
 # -*- coding: utf-8 -*-
-from collections import OrderedDict
-import cherrypy
 import pymongo
-import six
-import re
 
-from ..constants import GIRDER_ROUTE_ID, SettingDefault, SettingKey
 from .model_base import Model
 from girder import logprint
 from girder.exceptions import ValidationException
-from girder.utility import config, mail_utils, setting_utilities
+from girder.settings import SettingDefault
+from girder.utility import setting_utilities
 from girder.utility._cache import cache
-from bson.objectid import ObjectId
 
 
 class Setting(Model):
@@ -161,277 +156,3 @@ class Setting(Model):
             if callable(fn):
                 return fn()
         return None
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.BRAND_NAME)
-    def validateCoreBrandName(doc):
-        if not doc['value']:
-            raise ValidationException('The brand name may not be empty', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.BANNER_COLOR)
-    def validateCoreBannerColor(doc):
-        if not doc['value']:
-            raise ValidationException('The banner color may not be empty', 'value')
-        elif not (re.match(r'^#[0-9A-Fa-f]{6}$', doc['value'])):
-            raise ValidationException('The banner color must be a hex color triplet', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.PRIVACY_NOTICE)
-    def validateCorePrivacyNotice(doc):
-        if not doc['value']:
-            raise ValidationException('The privacy notice link may not be empty', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SECURE_COOKIE)
-    def validateSecureCookie(doc):
-        if not isinstance(doc['value'], bool):
-            raise ValidationException('Secure cookie option must be boolean.', 'value')
-
-    @staticmethod
-    @setting_utilities.default(SettingKey.SECURE_COOKIE)
-    def defaultSecureCookie():
-        return config.getConfig()['server']['mode'] == 'production'
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.ADD_TO_GROUP_POLICY)
-    def validateCoreAddToGroupPolicy(doc):
-        doc['value'] = doc['value'].lower()
-        if doc['value'] not in ('never', 'noadmin', 'nomod', 'yesadmin', 'yesmod', ''):
-            raise ValidationException(
-                'Add to group policy must be one of "never", "noadmin", '
-                '"nomod", "yesadmin", or "yesmod".', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.COLLECTION_CREATE_POLICY)
-    def validateCoreCollectionCreatePolicy(doc):
-        from .group import Group
-        from .user import User
-
-        value = doc['value']
-
-        if not isinstance(value, dict):
-            raise ValidationException('Collection creation policy must be a JSON object.')
-
-        for i, groupId in enumerate(value.get('groups', ())):
-            Group().load(groupId, force=True, exc=True)
-            value['groups'][i] = ObjectId(value['groups'][i])
-
-        for i, userId in enumerate(value.get('users', ())):
-            User().load(userId, force=True, exc=True)
-            value['users'][i] = ObjectId(value['users'][i])
-
-        value['open'] = value.get('open', False)
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.COOKIE_LIFETIME)
-    def validateCoreCookieLifetime(doc):
-        try:
-            doc['value'] = int(doc['value'])
-            if doc['value'] > 0:
-                return
-        except ValueError:
-            pass  # We want to raise the ValidationException
-        raise ValidationException('Cookie lifetime must be an integer > 0.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SERVER_ROOT)
-    def validateCoreServerRoot(doc):
-        val = doc['value']
-        if val and not val.startswith('http://') and not val.startswith('https://'):
-            raise ValidationException('Server root must start with http:// or https://.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.CORS_ALLOW_METHODS)
-    def validateCoreCorsAllowMethods(doc):
-        if isinstance(doc['value'], six.string_types):
-            methods = doc['value'].replace(',', ' ').strip().upper().split()
-            # remove duplicates
-            methods = list(OrderedDict.fromkeys(methods))
-            doc['value'] = ', '.join(methods)
-            return
-        raise ValidationException(
-            'Allowed methods must be a comma-separated list or an empty string.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.CORS_ALLOW_HEADERS)
-    def validateCoreCorsAllowHeaders(doc):
-        if isinstance(doc['value'], six.string_types):
-            headers = doc['value'].replace(",", " ").strip().split()
-            # remove duplicates
-            headers = list(OrderedDict.fromkeys(headers))
-            doc['value'] = ", ".join(headers)
-            return
-        raise ValidationException(
-            'Allowed headers must be a comma-separated list or an empty string.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.CORS_ALLOW_ORIGIN)
-    def validateCoreCorsAllowOrigin(doc):
-        if isinstance(doc['value'], six.string_types):
-            origins = doc['value'].replace(",", " ").strip().split()
-            origins = [origin.rstrip('/') for origin in origins]
-            # remove duplicates
-            origins = list(OrderedDict.fromkeys(origins))
-            doc['value'] = ", ".join(origins)
-            return
-        raise ValidationException(
-            'Allowed origin must be a comma-separated list of base urls or * or an empty string.',
-            'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.CORS_EXPOSE_HEADERS)
-    def validateCoreCorsExposeHeaders(doc):
-        if not isinstance(doc['value'], six.string_types):
-            raise ValidationException('CORS exposed headers must be a string')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.EMAIL_FROM_ADDRESS)
-    def validateCoreEmailFromAddress(doc):
-        # mail_utils.validateEmailAddress cannot be used here, as RFC 5322 allows this to accept an
-        # an address which includes a display name too
-        if not doc['value']:
-            raise ValidationException('Email from address must not be blank.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.CONTACT_EMAIL_ADDRESS)
-    def validateCoreContactEmailAddress(doc):
-        # This is typically used within an RFC 6068 "mailto:" scheme, so no display name is allowed
-        if not mail_utils.validateEmailAddress(doc['value']):
-            raise ValidationException(
-                'Contact email address must be a valid email address.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.EMAIL_HOST)
-    def validateCoreEmailHost(doc):
-        if isinstance(doc['value'], six.string_types):
-            doc['value'] = doc['value'].strip()
-            return
-        raise ValidationException('Email host must be a string.', 'value')
-
-    @staticmethod
-    @setting_utilities.default(SettingKey.EMAIL_HOST)
-    def defaultCoreEmailHost():
-        if cherrypy.request and cherrypy.request.local and cherrypy.request.local.name:
-            host = '%s://%s' % (cherrypy.request.scheme, cherrypy.request.local.name)
-            if cherrypy.request.local.port != 80:
-                host += ':%d' % cherrypy.request.local.port
-            return host
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.REGISTRATION_POLICY)
-    def validateCoreRegistrationPolicy(doc):
-        doc['value'] = doc['value'].lower()
-        if doc['value'] not in ('open', 'closed', 'approve'):
-            raise ValidationException(
-                'Registration policy must be "open", "closed", or "approve".', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.EMAIL_VERIFICATION)
-    def validateCoreEmailVerification(doc):
-        doc['value'] = doc['value'].lower()
-        if doc['value'] not in ('required', 'optional', 'disabled'):
-            raise ValidationException(
-                'Email verification must be "required", "optional", or "disabled".', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.API_KEYS)
-    def validateApiKeys(doc):
-        if not isinstance(doc['value'], bool):
-            raise ValidationException('API key setting must be boolean.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.ENABLE_PASSWORD_LOGIN)
-    def validateEnablePasswordLogin(doc):
-        if not isinstance(doc['value'], bool):
-            raise ValidationException('Enable password login setting must be boolean.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.ROUTE_TABLE)
-    def validateCoreRouteTable(doc):
-        nonEmptyRoutes = [route for route in doc['value'].values() if route]
-        if GIRDER_ROUTE_ID not in doc['value'] or not doc['value'][GIRDER_ROUTE_ID]:
-            raise ValidationException('Girder root must be routable.')
-
-        for key in doc['value']:
-            if (doc['value'][key] and not doc['value'][key].startswith('/')):
-                raise ValidationException('Routes must begin with a forward slash.')
-
-        if len(nonEmptyRoutes) > len(set(nonEmptyRoutes)):
-            raise ValidationException('Routes must be unique.')
-
-    @staticmethod
-    @setting_utilities.default(SettingKey.ROUTE_TABLE)
-    def defaultCoreRouteTable():
-        return {
-            GIRDER_ROUTE_ID: '/'
-        }
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SMTP_HOST)
-    def validateCoreSmtpHost(doc):
-        if not doc['value']:
-            raise ValidationException('SMTP host must not be blank.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SMTP_PORT)
-    def validateCoreSmtpPort(doc):
-        try:
-            doc['value'] = int(doc['value'])
-            if doc['value'] > 0:
-                return
-        except ValueError:
-            pass  # We want to raise the ValidationException
-        raise ValidationException('SMTP port must be an integer > 0.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SMTP_ENCRYPTION)
-    def validateCoreSmtpEncryption(doc):
-        if not doc['value'] in ['none', 'starttls', 'ssl']:
-            raise ValidationException(
-                'SMTP encryption must be one of "none", "starttls", or "ssl".', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SMTP_USERNAME)
-    def validateCoreSmtpUsername(doc):
-        # any string is acceptable
-        pass
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.SMTP_PASSWORD)
-    def validateCoreSmtpPassword(doc):
-        # any string is acceptable
-        pass
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.UPLOAD_MINIMUM_CHUNK_SIZE)
-    def validateCoreUploadMinimumChunkSize(doc):
-        try:
-            doc['value'] = int(doc['value'])
-            if doc['value'] >= 0:
-                return
-        except ValueError:
-            pass  # We want to raise the ValidationException
-        raise ValidationException('Upload minimum chunk size must be an integer >= 0.', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.USER_DEFAULT_FOLDERS)
-    def validateCoreUserDefaultFolders(doc):
-        if doc['value'] not in ('public_private', 'none'):
-            raise ValidationException(
-                'User default folders must be either "public_private" or "none".', 'value')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.GIRDER_MOUNT_INFORMATION)
-    def validateCoreGirderMountInformation(doc):
-        value = doc['value']
-        if not isinstance(value, dict) or 'path' not in value:
-            raise ValidationException(
-                'Girder mount information must be a dict with the "path" key.')
-
-    @staticmethod
-    @setting_utilities.validator(SettingKey.ENABLE_NOTIFICATION_STREAM)
-    def validateEnableNotificationStream(doc):
-        if not isinstance(doc['value'], bool):
-            raise ValidationException(
-                'Enable notification stream option must be boolean.', 'value')
